+++ src/databricks/azext_databricks/aaz/latest/databricks/access_connector/_create.py
@@ -19,6 +19,9 @@ class Create(AAZCommand):
 
     :example: Create a databricks accessConnector
         az databricks access-connector create --resource-group MyResourceGroup --name MyAccessConnector --location westus --identity-type SystemAssigned
+
+    :example: Create a databricks accessConnector with identities
+        az databricks access-connector create --resource-group MyResourceGroup --name MyAccessConnector --location westus --identity-type UserAssigned --user-assigned-identities {"/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ManagedIdentity/userAssignedIdentities/{identityName}":{}}
     """
 
     _aaz_info = {