+++ src/databricks/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 +++++
 * Upgrade API version from 2022-04-01-preview to 2022-10-01-preview and 2023-02-01
 * az databricks workspace update: add --sku to change the sku tier name
+* az databricks access-connector create/update: add --identities to set the user assigned identities associated with the resource.
 
 0.9.0
 +++++